+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 1.1 (YYYY-MM-DD)
 ----------------
 * Add general support for **Python 2.7**.
+* All numeric methods are now supported.
 * Make ``__nonzero__()`` and ``__unicode__()`` exlusive to Python 2.x.
 * Make ``__bool__()`` and ``__bytes__()`` exclusive to Python 3.x.
 * Make ``__length_hint()`` exclusive to Python 3.4.